+++ app/src/main/java/com/fpf/smartscan/ui/components/concepts/ConceptItemsList.kt
@@ -34,7 +34,7 @@ import androidx.compose.ui.unit.dp
 import androidx.paging.compose.LazyPagingItems
 import coil3.compose.AsyncImagePainter
 import com.fpf.smartscan.media.MediaItem
-import com.fpf.smartscan.ui.components.media.MediaDescriptionCard
+import com.fpf.smartscan.ui.components.media.ConceptMediaItemCard
 import kotlinx.coroutines.launch
 import kotlin.math.roundToInt
 
@@ -122,7 +122,7 @@ fun ConceptItemsList(
                 val item = items[index]
 
                 if (item != null) {
-                    MediaDescriptionCard(
+                    ConceptMediaItemCard(
                         item=item,
                         onItemClick=onItemClick,
                         onItemLongClick = onItemLongClick,
